+++ src/assets/images/index.ts
@@ -1,0 +1,13 @@
+import kitty from "./catwobg4.png";
+import dailyCipher from "./daily-cipher.png";
+import dailyCombo from "./daily-combo.png";
+import dailyReward from "./daily-reward.png";
+import dollarCoin from "./dollar-coin.png";
+
+export {
+    dailyCipher,
+    dailyCombo,
+    dailyReward,
+    dollarCoin,
+    kitty
+};